safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFEOffsetElement-svgdom-in-prop-expected.txt

CONSOLE MESSAGE: line 45: TypeError: 'undefined' is not an object (evaluating 'offsetElement.in1.baseVal = "SourceGraphic"')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL offsetElement.in1.baseVal should be blur. Threw exception TypeError: 'undefined' is not an object (evaluating 'offsetElement.in1.baseVal')
